+++ changelog.md
@@ -47,6 +47,7 @@ __                                                                              
 ### Others
 * **Web Workers!!** Now The Game Run 2 times faster in campaing mode
 * Bots Now Start With Items
+* Remove Expanded Inventory
 __                                                                                                      __
 ## Fixes
 * **REQUEST ANIMATION FRAME!!!!!** No more Lag